[IE-164] 워크스페이스 툴바 리팩토링
