--- apps/client/src/components/WorkspaceView/components/WorkspaceToolbar/hooks/useAddNamespaceToolbarButtonProps.ts
@@ -1,15 +1,0 @@
-import { ToolbarButtonProps } from '@components/Toolbar/components/ToolbarButton';
-import useAddNamespaceClickHandler from '@hooks/namespace/useAddNamespaceClickHandler';
-import AddIcon from '@mui/icons-material/Add';
-
-function useAddNamespaceToolbarButtonProps(): ToolbarButtonProps {
-  const addNamespace = useAddNamespaceClickHandler();
-
-  return {
-    IconComponent: AddIcon,
-    label: '네임스페이스 추가',
-    onClick: addNamespace,
-  };
-}
-
-export default useAddNamespaceToolbarButtonProps;
