+++ apps/client/src/components/WorkspaceView/components/WorkspaceToolbar/index.tsx
@@ -1,8 +1,11 @@
 import Toolbar from '@components/Toolbar';
-import { ToolbarButtonProps } from '@components/Toolbar/components/ToolbarButton';
-import useAddNamespaceToolbarButtonProps from '@components/WorkspaceView/components/WorkspaceToolbar/hooks/useAddNamespaceToolbarButtonProps';
-import useChangeDirectoryToolbarButtonProps from '@components/WorkspaceView/components/WorkspaceToolbar/hooks/useChangeDirectoryToolbarButtonProps';
-import useOpenNativeFileManagerToolbarButtonProps from '@components/WorkspaceView/components/WorkspaceToolbar/hooks/useOpenNativeFileManagerToolbarButtonProps';
+import useChangeDirectoryClickHandler from '@components/WorkspaceView/components/WorkspaceToolbar/hooks/useChangeDirectoryClickHandler';
+import useOpenNativeFileManagerClickHandler from '@components/WorkspaceView/components/WorkspaceToolbar/hooks/useOpenNativeFileManagerClickHandler';
+import useAddNamespaceClickHandler from '@hooks/namespace/useAddNamespaceClickHandler';
+import useFileManagerName from '@hooks/useFileManagerName';
+import FolderCopyOutlinedIcon from '@mui/icons-material/FolderCopyOutlined';
+import OpenInNewIcon from '@mui/icons-material/OpenInNew';
+import PostAddIcon from '@mui/icons-material/PostAdd';
 import { useTheme } from '@mui/material';
 
 function WorkspaceToolbar() {
@@ -10,11 +13,11 @@ function WorkspaceToolbar() {
     palette: { divider },
   } = useTheme();
 
-  const toolbars: ToolbarButtonProps[] = [
-    useChangeDirectoryToolbarButtonProps(),
-    useOpenNativeFileManagerToolbarButtonProps(),
-    useAddNamespaceToolbarButtonProps(),
-  ];
+  const handleClickChangeDirectory = useChangeDirectoryClickHandler();
+  const handleClickOpenNativeFileManager = useOpenNativeFileManagerClickHandler();
+  const handleClickAddNamespace = useAddNamespaceClickHandler();
+
+  const fileManagerName = useFileManagerName();
 
   return (
     <Toolbar
@@ -22,7 +25,23 @@ function WorkspaceToolbar() {
         borderBottom: `1px solid ${divider}`,
         padding: '4px 32px',
       }}
-      buttons={toolbars}
+      buttons={[
+        {
+          IconComponent: FolderCopyOutlinedIcon,
+          label: '디렉토리 변경',
+          onClick: handleClickChangeDirectory,
+        },
+        {
+          IconComponent: OpenInNewIcon,
+          label: `${fileManagerName}에서 열기`,
+          onClick: handleClickOpenNativeFileManager,
+        },
+        {
+          IconComponent: PostAddIcon,
+          label: '네임스페이스 추가',
+          onClick: handleClickAddNamespace,
+        },
+      ]}
     />
   );
 }
